refactor: index components refactor
