+++ apps/extension/src/containers/playground/containers/accounts/accounts-home/accounts-home.css.ts
@@ -3,44 +3,6 @@ import { responsiveStyle } from 'ui/src/components-v2/system/theme-utils'
 import { vars } from 'ui/src/components-v2/system/theme.css'
 import { style, globalStyle } from '@vanilla-extract/css'
 
-export const accountIndexWrapper = style([
-	sprinkles({
-		background: 'backgroundSecondary',
-		position: 'sticky',
-		top: 0,
-		zIndex: 1,
-		display: 'flex',
-		paddingX: 'xlarge',
-		paddingTop: 'xlarge',
-		paddingBottom: 'large',
-		transition: 'slowall',
-	}),
-	{
-		// border: '1px solid red',
-		// transition: 'all 0.3s ease-out',
-	},
-])
-
-export const accountIndexWrapperShadow = style([
-	sprinkles({
-		paddingTop: 'medium',
-		paddingBottom: 'medium',
-	}),
-	{
-		boxShadow: '0px 10px 15px -3px rgba(0,0,0,0.1)',
-		paddingBottom: vars.spacing.small,
-	},
-])
-
-export const pricingText = style([
-	sprinkles({
-		// transition: 'slowall',
-	}),
-	{
-		// transition: 'font-size .3s ease',
-	},
-])
-
 export const panelWrapper = style([
 	sprinkles({
 		position: 'relative',
@@ -92,148 +54,6 @@ export const rightPanelAssetType = style([
 		// TODO: conditional align self for tokens pages
 		alignSelf: 'flex-start',
 		// maxHeight: '100%',
-	},
-])
-
-export const indexAssetsWrapper = style([
-	sprinkles({
-		position: 'relative',
-		marginX: 'xlarge',
-	}),
-	{},
-])
-
-export const indexAssetWrapper = style([
-	sprinkles({
-		position: 'relative',
-	}),
-	{},
-])
-
-export const indexAssetLinkRow = style([
-	sprinkles({
-		width: 'full',
-		color: 'borderDivider',
-		transition: 'fast',
-	}),
-	{
-		boxShadow: '0 -1px 0 0',
-		'::before': {
-			content: '""',
-			position: 'absolute',
-			color: 'borderDivider',
-			opacity: 0,
-			transition: vars.transition.fast,
-			top: 0,
-			bottom: 0,
-			left: `calc(${vars.spacing.medium} * -1)`,
-			right: `calc(${vars.spacing.medium} * -1)`,
-			pointerEvents: 'none',
-			background: vars.color.lead400,
-			borderRadius: vars.border.radius.medium,
-			boxShadow:
-				'0px 0px 0px 1px rgba(255, 255, 255, 0.15), 0px 136px 192px rgba(0, 0, 0, 0.3), 0px 50px 50px rgba(0, 0, 0, 0.25), 0px 24px 24px rgba(0, 0, 0, 0.2), 0px 12px 12px rgba(0, 0, 0, 0.15)',
-		},
-		selectors: {
-			'&:hover': {
-				// boxShadow: 'none',
-			},
-			'&:hover::before': {
-				// opacity: 1,
-			},
-		},
-	},
-])
-
-export const indexAssetLinkRowHover = style([
-	sprinkles({}),
-	{
-		boxShadow: 'none',
-		'::before': {
-			opacity: 1,
-		},
-	},
-])
-
-export const indexAssetLinkRowInner = style([
-	sprinkles({
-		position: 'relative',
-		flexGrow: 1,
-	}),
-	{
-		paddingTop: '24px',
-		paddingBottom: '24px',
-	},
-	responsiveStyle({
-		// mobile: { width: '100%' },
-		// tablet: { width: '33%' },
-		// desktop: { width: '25%' },
-	}),
-])
-
-export const teststyle = style([
-	sprinkles({
-		position: 'relative',
-		background: {
-			hover: 'red800',
-			focus: 'red900',
-		},
-	}),
-	{
-		border: '1px solid red',
-		'@media': {
-			[`screen and (min-width: 480px)`]: {
-				flexBasis: '50%',
-			},
-		},
-	},
-	responsiveStyle({
-		mobile: { width: '100%' },
-		tablet: { width: '33%' },
-		desktop: { width: '25%' },
-	}),
-])
-
-globalStyle(`.${darkMode} ${indexAssetLinkRow} > a:hover`, {
-	background: vars.color.wax500,
-})
-
-globalStyle(`${indexAssetWrapper}:first-child > a`, {
-	boxShadow: 'none',
-})
-
-export const indexAssetRowOverlay = style([
-	sprinkles({
-		position: 'absolute',
-		top: 0,
-		right: 0,
-		height: 'full',
-		display: 'flex',
-		alignItems: 'center',
-		pointerEvents: 'none',
-	}),
-	{},
-])
-
-export const indexAssetCircle = style([
-	sprinkles({
-		position: 'relative',
-		borderRadius: 'full',
-		pointerEvents: 'auto',
-		borderColor: 'backgroundSecondary',
-		borderStyle: 'solid',
-		borderWidth: 'xsmall',
-		transition: 'fast',
-	}),
-	{
-		width: '40px',
-		height: '40px',
-		marginLeft: '-9px',
-		selectors: {
-			'&:hover': {
-				borderColor: vars.color.purple500,
-			},
-		},
 	},
 ])
